requirements.txt
tensorflow
numpy
matplotlib
seaborn
scikit-learn
pillow
opencv-python
tk